Added RDCL.EVT_KEYS command
This is rudimentary and allows a user to query and return all event
keys/UIDs at once instead of having to scan through RDCL.EVT_QUERY where
the results may change part way through and is more taxing.

The danger with this command is that is blocking and may be akin to
swallowing/regurgitating a potato of data in the response.